requirements.txt
Django==4.2.2
djangorestframework==3.15.0
requests==2.32.3
django-phonenumber-field==8.0.0
phonenumberslite==8.13.40
django-cors-headers==4.4.0
pyTelegramBotAPI==4.21.0